% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{add.expt.to.database}
\alias{add.expt.to.database}
\title{Add uploaded data to the existing data base}
\usage{
add.expt.to.database(database.name, expt.info, prot.data, frac.data, std.data,
  origin.data, prefractionation.data, msmethod.data, dataproc.data)
}
\arguments{
\item{database.name}{path to the existing data base, e.g., proteomics.db}

\item{expt.info}{metadata associated with the experiment
(one row per expt) (still to come), e.g. data.frame(expt_id = input$expt.id, organism = input$organism)}

\item{prot.data}{data per protein per expt 
(incl all columns that are not frac.data)}

\item{frac.data}{data per protein per fraction per expt (here, mq.data)}

\item{std.data}{data per standard per fraction per expt (from mq.data)}

\item{origin.data}{required metadata about the specific experiment}

\item{prefractionation.data}{optional metatdata about the specific experiment, can be NULL}

\item{msmethod.data}{optional metatdata about the specific experiment, can be NULL}

\item{dataproc.data}{optional metatdata about the specific experiment, can be NULL}
}
\description{
Adds data from a new MaxQuant output file to the data base.
}
\details{
Usage example:
#origin_df <- data.frame(expt_id = expt.id, experimenter = "myself",
 #                       genotype = "unknown", cell_type = "yeast_cells",
 #                      harvest_date = "Nov 2016", buffer_composition = "TrisHCl", 
 #                      lysis_method = "standard", digestion_enzyme = "Trypsin", 
 #                      notes = NA)
 # msmethods_df <- data.frame(expt_id = expt.id, instrument_id = "X0000",
   #                       run_date = "Oct 2016", method_length = 1)
# dataproc_df <- data.frame(expt_id = expt.id, processing_platform = "unknown", 
   #                      search_algorithm = "unknown", filtering_algorithm = "unknown",
   #                      filtering_stringency = "unknown")
# prefractionation_df <- data.frame(expt_id = expt.id, column_id = "x",
 #                                amount_protein_loaded = 1, sample_vol_loaded = 1, 
 #                                lc_flow_rate = 1, lc_fraction_size = 1, 
 #                                time_per_fraction = 1, fractions_collected = 1)

## add the data to the db
# add.expt.to.database(database.name, data.frame(expt_id = expt.id, organism = "yeast"), 
#                     prot.data = NULL, frac.data = x, std.data = y, origin.data = origin_df, 
#                     prefractionation.data = prefractionation_df, 
#                     msmethod.data = msmethods_df, dataproc.data = dataproc_df)
}

