% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap.R
\name{aggregate_dt}
\alias{aggregate_dt}
\title{Aggregating replicate values}
\usage{
aggregate_dt(in.dt, measurement, agg_by)
}
\arguments{
\item{in.dt}{tidy data.table}

\item{measurement}{string indicating the type of measurment that will be used
for the aggregation. must be a column header in in.dt, e.g. "superSmu"}

\item{agg_by}{strings indicating the combination of column headers that identify
unique proteins, e.g., c("condition","replicate","gene_symbol")}
}
\value{
data.table with aggregated values
}
\description{
Aggregating replicate values
}
\details{
# summed.dt <- aggregate_dt(hmdat, measurement = "superSmu",agg_by = uniquifiers)
}
