% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_columns}
\alias{check_columns}
\title{Check for presence of named columns}
\usage{
check_columns(check.vector, input, input.name, function.name)
}
\arguments{
\item{check.vector}{String(s) to check, e.g. c("peaks", "gauss.width")}

\item{input}{The names() of this will be checked.}

\item{input.name}{Name of the input, e.g. "start.params"}

\item{function.name}{Name of function for which this test is carried out.}
}
\value{
Returns an error and stop signal if entries of check.vector are missing in the input data.
}
\description{
Check for presence of named columns
}
\details{
example usage: check_columns(c(what, color.by, split.by, y.lab, x, y), long.df, "long.df", "plot_profile")
}
