% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{check_nomenclature}
\alias{check_nomenclature}
\title{Check whether a supplied ID conforms with the nomenclatures we're using}
\usage{
check_nomenclature(ID = spikeIn)
}
\arguments{
\item{ID}{The UniProt or yeast gene ID to be checked.}
}
\value{
Returns a Boolean that indicates whether the supplied ID meets the 
criteria of either the UniProt or the yeast gene nomenclature.
}
\description{
Check whether a supplied ID conforms with the nomenclatures we're using
}

