% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{cleaning_MQ}
\alias{cleaning_MQ}
\title{Cleaning MaxQuant output}
\usage{
cleaning_MQ(mq.df, remove.contaminants = TRUE, remove.decoys = TRUE,
  poi = NULL, spikeIn = NULL)
}
\arguments{
\item{mq.df}{Data.frame that contains the raw MaxQuant results.
(see \code{\link{reading_MQ}})}

\item{remove.contaminants}{Boolean to indicate whether rows with "CONT"
should be removed (should be set to FALSE if non-yeast proteins are of 
interest, including spike-ins)}

\item{remove.decoys}{Boolean to indicate whether rows with "REV" 
should be removed  (should always be set to TRUE)}

\item{poi}{Specify the organism for which the proteins of interest should be
retrieved; currently that can be:
\itemize{
\item "human": only those hits where the  protein ID conforms with the 
TrEMBL nomenclature
\item "yeast": only those hits where the protein ID conforms with the
 yeast standard nomenclature of either starting with Y or Q
 }
 Note that `poi` is mutually exclusive with `spikeIn`, i.e., either one
 should be set to NULL.}

\item{spikeIn}{Either "trypsin" or a string with a single UniProtID of
interest, e.g. "P00761". Choosing "trypsin" will retrieve values for trypsin 
from pig (UniProt ID P00761). This option is mutually exclusive with `poi`, 
which should be set to NULL.}
}
\value{
A subset of the original MaxQuant data.frame.
}
\description{
Before generating a long data.frame, it is helpful to clean the MaxQuant 
output, i.e., to remove contaminants, decoy hits and other hits that do not
follow the nomenclature of annotated proteins.
There is a hierarchy to the parameters, i.e., contaminants are removed first,
followed by the removal of decoy hits etc.
}
\seealso{
\code{\link{MQ_to_longFormat}}, \code{\link{extract_proteinID}}
}
