% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap.R
\name{corrHM_prep}
\alias{corrHM_prep}
\title{Preparing tidy correlation data.table for heatmap visualization}
\usage{
corrHM_prep(in.dt)
}
\arguments{
\item{in.dt}{data.table in tidy format with:, e.g. output from pw_corr_map()}
}
\value{
matrix fit for most heatmap functions c("gene_symbol","comparison","correlation"),
i.e.: proteins per row, samples per columns, 
gene symbols as row names. If there's only one column, the matrix will be sorted in descending order.
}
\description{
Preparing tidy correlation data.table for heatmap visualization
}
