% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap.R
\name{corrHM_wrap}
\alias{corrHM_wrap}
\title{Wrapper for generating matrices of correlation coefficients}
\usage{
corrHM_wrap(in.dt, measurement = "superSmu", corr.method = "pearson",
  uniquifiers = uniquifiers, calc.mode = c("both", "pairwise", "condition"))
}
\arguments{
\item{in.dt}{data.table with one row per protein fraction and different
columns for different types of measurements}

\item{measurement}{the type of measurement that should be used for the correlation
calculation, e.g. "raw" or "superSmu"}

\item{corr.method}{string indicating which method should be used by the base R cor() function.
One of "pearson" (default), "kendall", or "spearman"}

\item{uniquifiers}{vector of strings that define a unique sample,
e.g.: c("gene_symbol","condition","replicate")}

\item{calc.mode}{indicate whether you want to calculate the correlation for each pair of samples ["pairwise"], 
or between conditions ["condition"], or both ["both"].}
}
\value{
List with one or two matrices for heatmap visualization.
}
\description{
Wrapper for generating matrices of correlation coefficients
}
\seealso{
\code{\link{cor}}
}

