% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{extract_proteinID}
\alias{extract_proteinID}
\title{Extract ID information}
\usage{
extract_proteinID(prot.id, routine = NULL, regex = NULL, label = NULL)
}
\arguments{
\item{prot.id}{Vector of protein.id entries or the name of the column 
with related information}

\item{routine}{Specify the type of IDs you would like to retrieve.
Currently, the following options are available:
 \itemize{
 \item "trypsin": retrieving the protein ID for pig trypsin
 \item "yeast": retrieves the protein IDs etc. that conform with the official 
 yeast nomenclature (either starting with Y or Q)
 \item "human": retrieves protein IDs that conform with the TrEMBL format 
 and mnemonic protein names of UniProt
 }}

\item{regex}{If you do not want to use one of the pre-established routines,
you should specify a **vector of two strings**, where the first string 
corresponds to the pattern and the second string to the replacement.
For example, regex =  c("F", "f") will replace
upper case F with lower case f within the protein ID column.}

\item{label}{If you don't want to use a routine, indicate a label for your 
customized filtering, e.g. "gene_IDs"}
}
\value{
If a routine is chosen, this returns a list of ID vectors. If regex
and label are used, the list will contain only one vector of IDs.
Both results can directly be used with
\code{\link{MQaccessor_to_plotname()}}, which is used within
\code{\link{MQ_to_longFormat()}}.
}
\description{
This function currently contains all the regex we've come up with to extract
IDs of interest as defined by the Dephoure lab.
}
\details{
To parse the protein-centric information into the data base using a consistent
ID scheme, these information must be extracted from the sometimes erratic
MaxQuant output.
Older MaxQuant formats have the information in $Fasta.headers, newer ones will
have a $Gene.names column and an empty $Fasta.headers column. 
We will therefore rely on the $Protein.IDs only and retrieve the gene_symbol
info from our external data base.
}
\seealso{
\code{\link{MQ_to_longFormat}}
}
