% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{extract_protspec}
\alias{extract_protspec}
\title{Extract all information from the MaxQuant output that is
present only once per protein}
\usage{
extract_protspec(mq.df)
}
\arguments{
\item{mq.df}{data.frame based on MaxQuant output}
}
\value{
data.frame of protein-specific entries
}
\description{
The MaxQuant table contains values per fraction, which can be extracted via
\code{\link{MQ_to_longFormat}}, and entries that are present only once per
protein ID. These entries include, but are not limited to,
the Protein ID itself, (sometimes) the FASTA headers, numbers of proteins and 
peptides detected as well as oxidation and phosphorylation sites.
}
\details{
\bold{example usage:}

mq <- reading_MQ(filename = "../dephourelab/DepLab/inst/extdata/test_data/proteinGroups_100mM_new.txt")

prot.specific <- extract_protspec(mq)
}

