% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap.R
\name{get_cor_dt}
\alias{get_cor_dt}
\title{Turn the 2D-correlation matrix into a 1D-data table}
\usage{
get_cor_dt(cor.mat)
}
\arguments{
\item{cor.mat}{matrix of pairwise correlations, obtained via cor()}
}
\value{
a data.table with one column for the pairwise correlations,
one column describing the comparison
}
\details{
\bold{example usage:}
cors.dt <- lapply(unique(dat.test$id),
}

