% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{get_UniProt_info}
\alias{get_UniProt_info}
\title{Retrieve the UniProt information via the protein ID}
\usage{
get_UniProt_info(protein.ID)
}
\arguments{
\item{protein.ID}{String with the UniProt accessor, e.g. "P00761"}
}
\value{
A string indicating the protein ID together with the first line of 
the information of the .txt file provided by UniProt
(e.g. http://www.uniprot.org/uniprot/P07761.txt)
plus the link to the corresponding html page, which provides more
information and is easier to look at.
}
\description{
This function will use the UniProt ID to access UniProt's information
about that protein via their API.
}
\details{
\bold{example usage:}

get_UniProt_info("P00761")
}

