% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{MQ_to_cumulativePlot}
\alias{MQ_to_cumulativePlot}
\title{Preparing data for plotting of "metadata"}
\usage{
MQ_to_cumulativePlot(mq.df, value)
}
\arguments{
\item{mq.df}{data frame with MaxQuant data}

\item{value}{string indicating which sort of value to summarize; one of:
c("raw.intensity", "MS.MS.count", "peptides.count",
"unique.peptides.only", "razor.and.unique.peptides")}
}
\value{
data.table suitable for plotting with ggplot2
}
\description{
"Metadata" here simply refers to the summing up of all measurements
in one fraction across all available peptides/proteins.
}
\seealso{
\code{\link{plot_cumulativeValuesAcrossFractions}},
\code{\link{MQ_to_ggplot2}}
}
