% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{MQ_to_longFormat}
\alias{MQ_to_longFormat}
\title{Generate a long data.frame for a specific MS value}
\usage{
MQ_to_longFormat(mq.df, y, return.dt = FALSE, ...)
}
\arguments{
\item{mq.df}{data.frame based on MaxQuant output without 
decoy hits (REV) and contaminants (see also \cite{\link{cleaning_MQ}}}

\item{y}{type of value to be used for y-axis values}

\item{return.dt}{Boolean indicating whether a data.table or data.frame should be returned.
Default: FALSE}

\item{...}{several vectors (or a list thereof), e.g. of protein IDs,
that will be merged with the actual values. These should therefore have the
same length as the data.frame!}
}
\description{
One problem with the MaxQuant output is that there is a separate column
for each value/fraction pair.
This function will generate a long data.frame where each row contains the
value of interest per fraction.
}
\seealso{
\code{\link{reading_MQ}}, \code{\link{MQaccessor_to_plotname}}, 
\code{link{extract_proteinID}}, \code{\link{read.mq.data}}
}

