% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{MQaccessor_to_plotname}
\alias{MQaccessor_to_plotname}
\title{Match the MaxQuant accessor to the plot name}
\usage{
MQaccessor_to_plotname(plotname)
}
\arguments{
\item{plotname}{type of value to be used for y-axis values in the fraction
profile plot}
}
\value{
Regular expression for extracting all columns associated with the 
desired value of interest from the MaxQuant output.
}
\description{
This is a helper function that matches the y-axis value name that will
be displayed in the fraction profile plot with the regex needed 
for extracting the data from the MaxQuant output.
}
\details{
This function is needed because the MaxQuant output has a separate column 
for each value per fraction, but we need a long data.frame where the fraction
is indicated in an extra column.
This function helps to a) retrieve all the columns from the MaxQuant output
related to a column and b) to keep a meaningful name for the value that can
be used for the axis labeling in the plot later on.
}
