% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{norm_by_fraction}
\alias{norm_by_fraction}
\title{Normalize values across fractions}
\usage{
norm_by_fraction(long.dt)
}
\arguments{
\item{long.dt}{data.table with one row per protein and fraction; the keys
should be set to identify unique proteins and experiment IDs.
Fractions should not be key'ed!}
}
\value{
It returns virtually the same data.table except that the 'value'
column will contain normalized values and the 'measurement' column will have
been updated to reflect the normalization.
}
\description{
This normalization assumes that proteins present at equal levels in each
fraction should get a value of 1 in each fraction.
("row-wise normalization")
}
\seealso{
\code{\link{normalize_values}}
}

