% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cumulativeValues}
\alias{plot_cumulativeValues}
\title{Plotting "metadata"}
\usage{
plot_cumulativeValues(long.df, x.interval = 5, color.palette = NULL,
  title = NULL, split.by = NULL, y.lab = NULL)
}
\arguments{
\item{long.df}{data.frame or data.table fit for ggplot2, 
generated with MQ_to_cumulativePlot}

\item{x.interval}{number indicating the interval for labeling the fractions}

\item{color.palette}{name of RColorBrewer palette used for plotting; 
see http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/figure/
unnamed-chunk-14-1.png for palette names}

\item{title}{title for the plot}
}
\value{
A bar plot visualizing the sum of the value of interest across all
fractions.
}
\description{
For plotting the sum of certain values per fraction,
e.g., the sum of all unique peptides identified per fraction.
}
\seealso{
\code{\link{MQ_to_cumulativePlot}}
}
