% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap.R
\name{pw_corr_wrap}
\alias{pw_corr_wrap}
\title{Wrapper for calculating pairwise correlations}
\usage{
pw_corr_wrap(in.dt, gene, measurement, uniq.factors = c("gene_symbol",
  "fraction", "condition", "replicate"), cast.form, ...)
}
\arguments{
\item{in.dt}{data.table in tidy format}

\item{gene}{must match an entry in the gene_symbol column of the in.dt}

\item{uniq.factors}{vector of strings that determines how the pairwise
correlations will be calculated, e.g. c("gene_symbol","fraction","condition","replicate")}

\item{cast.form}{formula for casting the pairwise comparisons, in string format,
e.g., "fraction  ~  condition + replicate"}

\item{...}{additional parameters for \code{\link{cor}}}

\item{mesasurement}{string indicating the type of value that should be used for
the correlation calculation, e.g. "superSmu" or "raw"}
}
\value{
tidy data.table of pairwise correlations; if no correlation can be
determined (e.g. because there are not enough values), nothing will be returned
comparison correlation               gene_symbol
#1: D109N_01_vs_EV_01   0.7964631    A0AVT1 (UBA6)
#2: D109N_01_vs_EV_02   0.7659069    A0AVT1 (UBA6)
#3: D109N_01_vs_WT_01   0.8525967    A0AVT1 (UBA6)
#4: D109N_01_vs_WT_02   0.7876166    A0AVT1 (UBA6)
#5:    EV_01_vs_EV_02   0.9859125    A0AVT1 (UBA6)
#6:    EV_01_vs_WT_01   0.9897667    A0AVT1 (UBA6)
#7:    EV_01_vs_WT_02   0.9984263    A0AVT1 (UBA6)
#8:    EV_02_vs_WT_01   0.9587254    A0AVT1 (UBA6)
#9:    EV_02_vs_WT_02   0.9866967    A0AVT1 (UBA6)
#10:    WT_01_vs_WT_02   0.9865275    A0AVT1 (UBA6)
}
\description{
This function will calcule the pairwise correlation of multiple profiles for 
one specific protein or gene.
}
\details{
\bold{example usage:}
# cors.dt <- lapply(unique(dat$gene_symbol), function(x) pw_corr_wrap(in.dt = dat, gene = x, method = "pearson") ) %>% rbindlist
}
