% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{query.measurements.by.expt.v2}
\alias{query.measurements.by.expt.v2}
\title{Query the frac_data table}
\usage{
\method{query}{measurements.by.expt.v2}(database.name, expt.ids,
  measurement.type)
}
\description{
Filtering by expt_id, and measurement
}

