% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{query.measurements.by.expt.with.gene.symbol.v2}
\alias{query.measurements.by.expt.with.gene.symbol.v2}
\title{Query the frac_data table}
\usage{
\method{query}{measurements.by.expt.with.gene.symbol.v2}(database.name,
  expt.ids, measurement.type)
}
\description{
Query the data for fract data and provide gene symbols in output
}
