% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{query.measurements}
\alias{query.measurements}
\title{Query the frac_data table I}
\usage{
\method{query}{measurements}(database.name, expt.ids, gene.symbols,
  measurement.type)
}
\description{
Filtering by expt_id, gene_symbol and measurement.
}

