% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{query.measurements.v2}
\alias{query.measurements.v2}
\title{Query the frac_data table I (v2)}
\usage{
\method{query}{measurements.v2}(database.name, expt.ids, gene.symbols,
  measurement.type)
}
\arguments{
\item{database.name}{Path to data base.}

\item{expt.ids}{One or more experiment IDs for which the values should be retrieved.}

\item{gene.symbols}{The specific, unique gene symbols for which the values should be retrieved.}

\item{measurement.type}{One of c("peptides.count","unique.peptides.only","razor.and.unique.peptides", "sequence.coverage","raw.intensity").}
}
\description{
Filtering by expt_id, gene_symbol and measurement.
This version uses native sql instead of dplyr code.
}

