% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{query.std.measurements.by.expt}
\alias{query.std.measurements.by.expt}
\title{Query the std_data table}
\usage{
\method{query}{std.measurements.by.expt}(database.name, expt.ids,
  measurement.type)
}
\description{
Query the data for the spike-in while filtering by expt_id and measurement.
}
