% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{query.std.measurements.v2}
\alias{query.std.measurements.v2}
\title{Query the std_data table}
\usage{
\method{query}{std.measurements.v2}(database.name, expt.ids, std.ids,
  measurement.type)
}
\description{
Filtering by expt_id, gene_symbol and measurement.
}
