% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{read_complexes}
\alias{read_complexes}
\title{Read in lists of known complexes}
\usage{
read_complexes(filename, organism)
}
\arguments{
\item{filename}{Path to file. Should have one row per protein with at least:
'id', 'complex'}

\item{organism}{Organism for which the complex list was generated, e.g. 'human'.}
}
\value{
Data.frame with 'id', 'gene_symbol' and 'complex' where the gene symbol
corresponds to the gene symbol format used for all the raw data
( < UniProtID (gene name) > ).
}
\description{
This function reads in a table where for each protein a possible
complex membership is noted.
}

