% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{read_IDinfo}
\alias{read_IDinfo}
\title{Pre-populate info table with all the known protein/gene names}
\usage{
read_IDinfo(organism)
}
\arguments{
\item{organism}{string indicating the organism for which to retrieve the information.}
}
\value{
Data.table with two columns, 'id' and 'gene_symbol'; 'id' is keyed.
Note that 'gene_symbol' is a bit of a misnomer as it's a mash up of the ID 
and the primary gene name.
}
\description{
To initiate the id.info data base, this function retrieves the
information stored in the ID_lists folder of the DepLab package.
The lists should have two columns: 1) unique IDs, e.g. UniProt identifiers;
2) easily recognizable gene names, e.g. the primary gene name from UniProt.
These will be pasted, so that the gene name column will also be unique.
}
\details{
\bold{Yeast data} derived from
\url{http://downloads.yeastgenome.org/curation/chromosomal_feature/SGD_features.tab}
and generated using an awk routine.

\bold{Human} info downloaded from uniprot.org with the following search:
organism:human AND organism:"Homo sapiens (Human) [9606]" 
AND proteome:up000005640

cut -f 1-2 *tab | sed "1d" > human_uniprot.entry.name
}
\seealso{
\code{\link{make_gene_name}}
}
