% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{reading_MQ}
\alias{reading_MQ}
\title{Reading in MaxQuant output}
\usage{
reading_MQ(filename)
}
\arguments{
\item{filename}{path to MaxQuant output file}
}
\description{
This function expects the MaxQuant output "proteinGroups*.txt"
and returns a data frame.
}

