% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{return.gene.symbols.from.id}
\alias{return.gene.symbols.from.id}
\title{Convert ID to gene symbol}
\usage{
return.gene.symbols.from.id(database.name, id.name, organism.name)
}
\description{
Filtering by id and organism
}
