% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{smooth_for_plot}
\alias{smooth_for_plot}
\title{Smoothening the data for plotting}
\usage{
smooth_for_plot(in.df, what = c("gene_symbol", "expt_id"),
  x.values = "fraction", y.values = "value")
}
\arguments{
\item{in.df}{Data.frame}

\item{what}{indicate how the points that will be connected by lines should be
grouped together, e.g. c("gene_symbol", "expt_id") will draw a
separate line for each unique combination of gene_symbol and expt_id}

\item{x.values}{string indicating the column containing the values to be plotted
on the x.axis}

\item{y.values}{string indicating the column containing the values to be plotted on the y.axis}
}
\value{
Data frame with smoothened values.
}
\description{
This calculates X-spline values relative to control points.
}
\details{
The data.frame will be sorted first by factor2, then factor1, then the x.values.
The smoothened values will be calculated using ddplyr(in.df, factor2 ~ factor1).
}
\seealso{
\code{\link{xspline}}, \code{\link{plot_profile}}
}

