% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{superSmooth_values}
\alias{superSmooth_values}
\title{Smooth measured values}
\usage{
superSmooth_values(long.df, prot.identifier = NULL)
}
\arguments{
\item{long.df}{data.frame with unique values per protein, expt. ID, 
measurement and fraction; preferably for only one measurement.}

\item{prot.identifier}{specify the name that identifies single proteins,
e.g. "gene_symbol"}
}
\value{
data.frame with normalized intensity values
}
\description{
This function allows for the smoothening of, e.g., raw intensity values
collected from the MaxQuant output using Friedman's superSmoother.
}
\details{
\bold{example usage:}
}
\seealso{
\code{\link{supsmu}}
}

