% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{update.expt}
\alias{update.expt}
\title{Edit experiment meta-data}
\usage{
\method{update}{expt}(database.name, expt.ids, new.expt.id, new.organism.name)
}
\description{
Edit experiment meta-data, e.g. organism.
}
