% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{read.MQ.data}
\alias{read.MQ.data}
\title{Reading MaxQuant data}
\usage{
read.MQ.data(filename, expt.id, data.subset = "poi", organism = NULL,
  return.dt = FALSE, db.order.poi = c("gene_symbol", "fraction", "value",
  "measurement", "expt_id", "organism"), db.order.std = c("id", "fraction",
  "value", "measurement", "expt_id"))
}
\arguments{
\item{filename}{path to MaxQuant output (proteinGroups.txt)}

\item{expt.id}{unique identifier for the experiment to be read in}

\item{data.subset}{a string specifying what kind of data should be retrieved,
one of:
\itemize{
\item "poi": extract information for proteins of interest -- if you choose
this option, make sure to indicate the organism!
\item "trypsin": extract information for trypsin
\item a specific UniProt ID or yeast ORF, e.g. "P00761"
}}

\item{organism}{This determines how the proteins of interest should be
retrieved, i.e., indicate either 'human' or 'yeast', or NULL if data.subset 
is not 'poi'.}

\item{return.dt}{Boolean indicating whether a data.table shold be returned
instead of a data.frame. Default: FALSE}

\item{db.order.poi}{Vector indicating the _order_ of the columns for the
data.frame holding the values for the proteins of interest. This should 
match the expectations of \code{\link{initialize.datbase}} and
\code{\link{add.expt.to.database}}.}

\item{db.order.std}{Vector indicating the _order_ of the columns for the
data.frame holding the values for the spiked-in proteins. This should 
match the expectations of \code{\link{initialize.datbase}} and
\code{\link{add.expt.to.database}}.}
}
\value{
A long data.frame where each row corresponds to a single protein,
and one type of measured value per fraction.
}
\description{
This reads in the MaxQuant output, retrieves the proteins of interest and
generates a long data.frame for _all types of measurement_ that are available
across all fractions.
}
\seealso{
@seealso \code{\link{reading_MQ}}, \code{\link{cleaning_MQ}},
\code{\link{extract_proteinID}}
}
