% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{update.expt}
\alias{update.expt}
\title{Edit experiment metadata}
\usage{
\method{update}{expt}(database.name, expt.ids, new.expt.id, new.organism.name,
  origin.data, prefractionation.data, msmethod.data, dataproc.data)
}
\description{
Edit experiment metadata, e.g. organism.
}
